requirements.txt
# FastAPI
fastapi==0.110.0
uvicorn[standard]==0.23.1
gunicorn==21.2.0

# Redis async client
redis[async]==6.4.0

# PostgreSQL
databases[postgresql]
asyncpg
psycopg2-binary

# Neo4j
neo4j

# OpenAI
openai

# Bcrypt for password hashing
bcrypt

# dotenv
python-dotenv

# PDF generation
reportlab

# Optional: for data processing / embeddings
numpy
faiss-cpu
fastembed

# Pydantic
pydantic

# Requests for any HTTP calls
requests

